id의 Data Type Long -> Int

+++ app/src/main/java/com/example/jup_jup_android/entity/dataclass/ItemStatus.kt
@@ -1,14 +1,12 @@
 package com.example.jup_jup_android.entity.dataclass
 
-import android.graphics.Bitmap
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.io.Serializable
-import java.util.*
 
 @Entity(tableName = "itemStatus")
 data class ItemStatus(
-        @PrimaryKey var id: Long,
+        @PrimaryKey var id: Int,
         var image: String,
         var name: String,
         var category: String,
